+++ packages/backend/src/prompt/decision-loop-prompts.ts
@@ -1,6 +1,8 @@
 import { createPromptTemplate } from "@tambo-ai-cloud/core";
 
-export function generateDecisionLoopPrompt() {
+export function generateDecisionLoopPrompt(
+  customInstructions: string | undefined,
+) {
   return createPromptTemplate(
     `
 You are a friendly assistant that helps the user interact with an application.
@@ -25,7 +27,14 @@ For example, imagine these tools are available:
 
 If a user asks for weather in a city, you may call the 'get_weather' tool, and
 then call the 'show_component_Weather' tool to pass the weather information to the Weather component on screen.
-`,
-    {},
+
+{custom_instructions}`,
+    {
+      custom_instructions: customInstructions
+        ? `In addition to the above, please also follow these additional instructions:
+${customInstructions}
+`
+        : "",
+    },
   );
 }